moving to windows environment

+++ MyFirstBootedApp/Core/Src/crc32.c
@@ -7,7 +7,7 @@
 
 #include <crc32.h>
 
-uint32_t crc32(const uint8_t *src, uint32_t len, uint32_t state)
+uint32_t crc32(const uint8_t *src, uint8_t len, uint32_t state)
 {
 	static uint32_t crctab[256];
